Patch to fix compile warning on mac platforms.
git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@1411 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ test/testckpt.c
@@ -341,7 +341,7 @@ sleep (1);
 	timersub (&tv_end, &tv_start, &tv_elapsed);
 	printf ("Elapsed Time to expiry is %ld & %ld usec (should be about %d seconds)\n",
 		tv_elapsed.tv_sec,
-		tv_elapsed.tv_usec,
+		(long) tv_elapsed.tv_usec,
 		SECONDS_TO_EXPIRE);
 
 	error = saCkptCheckpointRetentionDurationSet (checkpointHandle,
